Revert Controller & Implement HTTP

+++ Controllers/GeneralController.cs
@@ -1,17 +1,20 @@
 ﻿using API.Contracts;
 using API.Data;
 using API.Models;
+using API.Utilities;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
 
 namespace API.Controllers;
 
 [ApiController]
-[Route("api/entities")]
-public class GeneralController<TEntity> : ControllerBase
+public class GeneralController<TRepository, TEntity> : ControllerBase
+    where TRepository : IGeneralRepository<TEntity>
+    where TEntity : class
 {
-    protected readonly IGeneralRepository<TEntity> _repository;
+    protected readonly TRepository _repository;
 
-    public GeneralController(IGeneralRepository<TEntity> repository)
+    public GeneralController(TRepository repository)
     {
         _repository = repository;
     }
@@ -23,10 +26,21 @@ public class GeneralController<TEntity> : ControllerBase
 
         if (!entities.Any())
         {
-            return NotFound();
+            return NotFound(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status404NotFound,
+                Status = HttpStatusCode.NotFound.ToString(),
+                Message = "Data not found!"
+            });
         }
 
-        return Ok(entities);
+        return Ok(new ResponseHandler<IEnumerable<TEntity>>
+        {
+            Code = StatusCodes.Status200OK,
+            Status = HttpStatusCode.OK.ToString(),
+            Message = "Data found",
+            Data = entities
+        });
     }
 
     [HttpGet("{guid}")]
@@ -36,42 +50,111 @@ public class GeneralController<TEntity> : ControllerBase
 
         if (entity is null)
         {
-            return NotFound();
+            return NotFound(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status404NotFound,
+                Status = HttpStatusCode.NotFound.ToString(),
+                Message = "Data not found!"
+            });
         }
 
-        return Ok(entity);
+        return Ok(new ResponseHandler<TEntity>
+        {
+            Code = StatusCodes.Status200OK,
+            Status = HttpStatusCode.OK.ToString(),
+            Message = "Data found"
+        });
     }
 
     [HttpPost]
     public IActionResult Create(TEntity entity)
     {
         var createdTEntity = _repository.Create(entity);
-        return Ok(createdTEntity);
+        if (createdTEntity is false)
+        {
+            return BadRequest(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status400BadRequest,
+                Status = HttpStatusCode.BadRequest.ToString(),
+                Message = "Data not created!"
+            });
+        }
+        return Ok(new ResponseHandler<TEntity>
+        {
+            Code = StatusCodes.Status200OK,
+            Status = HttpStatusCode.OK.ToString(),
+            Message = "Data created"
+        });
     }
 
     [HttpPut]
     public IActionResult Update(TEntity entity)
     {
+        var getGuid = (Guid)typeof(TEntity).GetProperty("Guid")!.GetValue(entity!);
+        var isFound = _repository.IsExist(getGuid);
+
+        if (isFound is false)
+        {
+            return NotFound(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status404NotFound,
+                Status = HttpStatusCode.NotFound.ToString(),
+                Message = "Id not found!"
+            });
+        }
+
         var isUpdated = _repository.Update(entity);
 
         if (!isUpdated)
         {
-            return NotFound();
+            return BadRequest(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status500InternalServerError,
+                Status = HttpStatusCode.InternalServerError.ToString(),
+                Message = "Check your data!"
+            });
         }
 
-        return Ok();
+        return Ok(new ResponseHandler<TEntity>
+        {
+            Code = StatusCodes.Status200OK,
+            Status = HttpStatusCode.OK.ToString(),
+            Message = "Successfully updated"
+        });
     }
 
     [HttpDelete("{guid}")]
     public IActionResult Delete(Guid guid)
     {
+        var isFound = _repository.IsExist(guid);
+
+        if (isFound is false)
+        {
+            return NotFound(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status404NotFound,
+                Status = HttpStatusCode.NotFound.ToString(),
+                Message = "Id not found"
+            });
+        }
+
         var isDeleted = _repository.Delete(guid);
 
         if (!isDeleted)
         {
-            return NotFound();
+            return BadRequest(new ResponseHandler<TEntity>
+            {
+                Code = StatusCodes.Status500InternalServerError,
+                Status = HttpStatusCode.InternalServerError.ToString(),
+                Message = "Check connection to database"
+            });
         }
 
-        return Ok();
+        return Ok(new ResponseHandler<TEntity>
+        {
+            Code = StatusCodes.Status200OK,
+            Status = HttpStatusCode.OK.ToString(),
+            Message = "Successfully deleted"
+        });
     }
 }
